extract questionably re-usable appendDefaultRow function
This seems like an awkward space where there is a meningful amount of code that would be reused, but the current API isn't amazing.

Seems worth trying later to make this into a stateful object and pack the classDefaultWrapper and classTotalOutput into it. Probably the entire preset controls while we're at it.